Sync main

+++ src/Augmentation/AugmentationHelpers.tsx
@@ -141,6 +141,12 @@ function applyAugmentation(aug: IPlayerOwnedAugmentation, reapply = false): void
     }
   }
 
+  // Special logic for Congruity Implant
+  if (aug.name === AugmentationNames.CongruityImplant && !reapply) {
+    Player.entropy = 0;
+    Player.applyEntropy(Player.entropy);
+  }
+
   // Push onto Player's Augmentation list
   if (!reapply) {
     const ownedAug = new PlayerOwnedAugmentation(aug.name);
@@ -148,8 +154,8 @@ function applyAugmentation(aug: IPlayerOwnedAugmentation, reapply = false): void
   }
 }
 
-function installAugmentations(): boolean {
-  if (Player.queuedAugmentations.length == 0) {
+function installAugmentations(force?: boolean): boolean {
+  if (Player.queuedAugmentations.length == 0 && !force) {
     dialogBoxCreate("You have not purchased any Augmentations to install!");
     return false;
   }
@@ -179,12 +185,14 @@ function installAugmentations(): boolean {
     augmentationList += aug.name + level + "<br>";
   }
   Player.queuedAugmentations = [];
-  dialogBoxCreate(
-    "You slowly drift to sleep as scientists put you under in order " +
-      "to install the following Augmentations:<br>" +
-      augmentationList +
-      "<br>You wake up in your home...you feel different...",
-  );
+  if (!force) {
+    dialogBoxCreate(
+      "You slowly drift to sleep as scientists put you under in order " +
+        "to install the following Augmentations:<br>" +
+        augmentationList +
+        "<br>You wake up in your home...you feel different...",
+    );
+  }
   prestigeAugmentation();
   return true;
 }
